kvcoord: merge truncate_test into batch_test
This commit moves `TestTruncate` as well as `BenchmarkTruncate` into
`batch_test.go` file in the preparation for the refactor done in the
following commit. The code itself wasn't changed at all.

Release note: None
